requirements.txt
streamlit
openai>=1.3.0
docx2txt>=0.8
pandas
requests
python-docxpython-docx>=0.8.11
